requirements.txt
asgiref==3.6.0
attrs==21.4.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
coverage==7.2.3
crispy-bootstrap4==2022.1
dj-database-url==1.2.0
Django==4.1.7
django-bootstrap4==22.3
django-crispy-forms==2.0
django-filter==23.2
exceptiongroup==1.0.4
flake8==5.0.4
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
requests-mock==1.10.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
validators==0.20.0
Werkzeug==2.2.2
whitenoise==6.4.0
